refactor(frontend): make chapter edit information form component
